Add files via upload
In this statistics project, we tried to implement various statistical techniques to prioritize the skill sets of a Data Science Professional. As a data sample, we conducted a survey on working professionals. The different job roles included are Data analyst, Data Scientist, Business Analyst and Data Engineer. We asked them to rate the different skills required as per level of importance on a scale of 1-5.
The techniques used for analysis is:
1.Using Estimation on each column to calculate the confidence interval of mean of each population with 95%confidence level.
2.Using Kruskalwallis Test to check whether any skill is statistically significant from the other for a particular job role. Post hoc pairwise test for multiple comparisons of mean rank sums (Dunn’s test). 
3.Using hypothesis testing to compare two populations to gather an idea if there is significant difference between both roles on the basis of mastering of the above skills.